Rename file_converters to converters
