+++ docs/pydoc/config-preview/converter.yml
@@ -1,6 +1,6 @@
 loaders:
   - type: loaders.CustomPythonLoader
-    search_path: [../../../haystack/preview/components/file_converters]
+    search_path: [../../../haystack/preview/components/converters]
     modules: ["azure", "html", "markdown", "pypdf", "tika", "txt"]
     ignore_when_discovered: ["__init__"]
 processors:
@@ -15,12 +15,12 @@ renderer:
   type: renderers.ReadmePreviewRenderer
   excerpt: Extracts text from files in different formats and converts it into the unified Document format.
   category_slug: haystack-classes
-  title: File Converter API
-  slug: file-converter-api
+  title: Converter API
+  slug: converter-api
   order: 50
   markdown:
     descriptive_class_title: false
     descriptive_module_title: true
     add_method_class_prefix: true
     add_member_class_prefix: false
-    filename: file_converter_api.md
+    filename: converter_api.md
