+++ haystack/preview/components/converters/__init__.py
@@ -1,0 +1,15 @@
+from haystack.preview.components.converters.txt import TextFileToDocument
+from haystack.preview.components.converters.tika import TikaDocumentConverter
+from haystack.preview.components.converters.azure import AzureOCRDocumentConverter
+from haystack.preview.components.converters.pypdf import PyPDFToDocument
+from haystack.preview.components.converters.html import HTMLToDocument
+from haystack.preview.components.converters.markdown import MarkdownToDocument
+
+__all__ = [
+    "TextFileToDocument",
+    "TikaDocumentConverter",
+    "AzureOCRDocumentConverter",
+    "PyPDFToDocument",
+    "HTMLToDocument",
+    "MarkdownToDocument",
+]
