--- haystack/preview/components/file_converters/__init__.py
@@ -1,15 +1,0 @@
-from haystack.preview.components.file_converters.txt import TextFileToDocument
-from haystack.preview.components.file_converters.tika import TikaDocumentConverter
-from haystack.preview.components.file_converters.azure import AzureOCRDocumentConverter
-from haystack.preview.components.file_converters.pypdf import PyPDFToDocument
-from haystack.preview.components.file_converters.html import HTMLToDocument
-from haystack.preview.components.file_converters.markdown import MarkdownToDocument
-
-__all__ = [
-    "TextFileToDocument",
-    "TikaDocumentConverter",
-    "AzureOCRDocumentConverter",
-    "PyPDFToDocument",
-    "HTMLToDocument",
-    "MarkdownToDocument",
-]
